refactor: import da url e remoção das functions
